Relocate middleware and force dynamic dashboard render
